Move the command
