+++ src/main/java/fr/xephi/authme/process/login/AsynchronousLogin.java
@@ -23,7 +23,6 @@ import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
 import org.bukkit.scheduler.BukkitTask;
 
-import java.util.Date;
 import java.util.List;
 
 /**
